deprstive by corolation is implimneted.

+++ SRC/mathOp.hxx
@@ -6,11 +6,11 @@ class mathOp
 {
 public:
 	mathOp();
-	void getVectors(vector<double> input_1, vector<double> input_2);
-	double calcMu(vector<double> input);
-	double calcCorro(vector<double> input_1, vector<double> input_2);
-	double calcVari(vector<double> input_1);
-	double calcVari(vector<double> input_1, double mu);
+	void getVectors(vector<int> input_1, vector<int> input_2);
+	double calcMu(vector<int> input);
+	double calcCorro(vector<int> input_1, vector<int> input_2);
+	double calcVari(vector<int> input_1);
+	double calcVari(vector<int> input_1, double mu);
 
 private:
 	double Mu_1;
@@ -18,8 +18,8 @@ private:
 	double Var_1;
 	double Var_2;
 
-	vector<double> input_1;
-	vector<double> input_2;
+	vector<int> input_1;
+	vector<int> input_2;
 
 };
